fix: moving waiter types to aws-sdk/types
